+++ src/AST.ml
@@ -87,6 +87,10 @@ type currency =
   | Eldritch_annul
   | Eldritch_exalt
   | Eldritch_chaos
+  | Wild_crystallised_lifeforce
+  | Vivid_crystallised_lifeforce
+  | Primal_crystallised_lifeforce
+  | Sacred_crystallised_lifeforce
   (* Not really currencies but... *)
   | Harvest_augment of Mod.harvest_tag
   | Harvest_non_to of Mod.harvest_tag
@@ -148,6 +152,10 @@ let show_currency = function
   | Eldritch_annul -> "eldritch_annul"
   | Eldritch_exalt -> "eldritch_exalt"
   | Eldritch_chaos -> "eldritch_chaos"
+  | Wild_crystallised_lifeforce -> "wild_lifeforce"
+  | Vivid_crystallised_lifeforce -> "vivid_lifeforce"
+  | Primal_crystallised_lifeforce -> "primal_lifeforce"
+  | Sacred_crystallised_lifeforce -> "sacred_lifeforce"
   | Harvest_augment tag -> "harvest_augment_" ^ Id.show (Mod.tag_id tag)
   | Harvest_non_to tag -> let t = Id.show (Mod.tag_id tag) in "harvest_non_" ^ t ^ "_to_" ^ t
   | Harvest_reforge tag -> "harvest_reforge_" ^ Id.show (Mod.tag_id tag)